replace use of :=> with new classes ForAnyKnownIndex
Using :=> worked, but it required weird instances; the form had to
always be KnownNat n :=> someConstraint, and if someConstraint actually
had other requirements those would have to be added as constraints on
the :=> instance. That means that any more usual use of :=> would have
had issues not being able to reify all of the real requirements.

And if I have to manually add a boilerplate instance for every instance
I want to be automatically lifted through Some, then I might as well use
my own class that says what it means:
ForAnyKnownIndex someConstraint f a says that someConstraint (f n a)
holds for any n, provided there is a KnownNat available for n.